chore: refactor to split the implementation to separated files
